Make catch HR properly utilise the RNG
